Implement Broadcast event auth and event handling
Only users who are members of a room can trigger events for their
respective Pusher channel (of the form: room.{room_id}).

+++ resources/js/components/RoomComponent.vue
@@ -24,9 +24,10 @@
                                                           v-bind:spotify-id="spotifyId"
                                                           v-bind:spotify-device-id="spotifyDeviceId"
                                                           v-bind:spotify-player-state="spotifyPlayerState"
-                                                          v-bind:playlist-id="playlistId">
+                                                          v-bind:playlist-id="playlistId"
                                                           v-bind:room-id="roomId"
-                                                          v-bind:has-broadcaster="hasBroadcaster">
+                                                          v-bind:has-broadcaster="hasBroadcaster"
+                                                          v-on:disconnect-session="disconnectSession">
                                 </spotify-player-component>
                                 <playlist-component v-bind:access-token="accessToken"
                                                     v-bind:spotify-id="spotifyId">
@@ -74,7 +75,6 @@
                 "spotifyPlayerState": {},
                 "playlistId": "",
                 "hasBroadcaster": false,
-                "isBroadcaster": false
             };
         },
         created() {
@@ -85,7 +85,7 @@
             this.initializeEventListeners();
             this.initializeSpotifyPlayer(this.currentAccessToken);
             this.getRoomBroadcasterStatus();
-            window.addEventListener("beforeunload", this.disconnectSession);
+            window.addEventListener("beforeunload", this.abruptlyCloseSession);
         },
         methods: {
             initializePlaylistId(token, id) {
@@ -112,11 +112,11 @@
             },
             initializeEventListeners() {
                 Echo.private(`room.${this.roomId}`)
-                    .listen("BroadcasterConnected", (eventObj) => {
-                        console.log("Broadcaster connected: " + eventObj);
+                    .listen("BroadcasterConnected", (data) => {
+                        this.hasBroadcaster = true;
                     })
-                    .listen("BroadcasterDisconnected", (eventObj) => {
-                        console.log("Broadcaster disconnected: " + eventObj);
+                    .listen("BroadcasterDisconnected", (data) => {
+                        this.hasBroadcaster = false;
                     });
             },
             initializeSpotifyPlayer(token) {
@@ -161,14 +161,20 @@
                     }
                 });
             },
-            disconnectSession() {
-                axios.delete("/api/room/" + this.roomId + "/broadcast")
-                .then((res) => {
-                    this.isBroadcaster = false;
-                })
-                .catch((err) => {
-                    console.log("Could not stop broadcasting.");
+            disconnectSession(isBroadcaster) {
+                spotifyApi.pause(function(err, data) {
+                    if (err) {
+                        console.error("Could not pause playback: " + err);
+                    }
                 });
+                if (isBroadcaster) {
+                    axios.delete('/api/room/' + this.roomId + '/broadcast');
+                    this.hasBroadcaster = false;
+                }
+            },
+            abruptlyCloseSession() {
+                spotifyApi.pause();
+                axios.delete('/api/room/' + this.roomId + '/broadcast');
             }
         },
     }
